% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-importer.R
\name{import_data_from_disk}
\alias{import_data_from_disk}
\title{Import Data from Disk}
\usage{
import_data_from_disk(data_source_path, data_target_path = NULL)
}
\arguments{
\item{data_source_path}{The path to the data source file on disk.}

\item{data_target_path}{The path where the data will be imported. Default is NULL,
which means the data will be imported to the current working directory.}
}
\description{
This function imports data from a specified data source path on disk to a target path.
If the target path is not provided, the function will use the current working directory.
}
\details{
The function checks if the data source file exists. If it doesn't exist, it provides appropriate
messages and returns without importing data.

If the data target path already exists, the function compares the hash of the source file with the
hash of the existing target file using MD5. If the hashes differ, the existing target file is removed,
and the data is copied from the source file to the target path, providing a message of data update.
If the hashes are the same, the function provides a message stating that the data is up-to-date.

If the data target path does not exist, the function copies the data from the source file to the target
path and provides a message indicating that the data has been imported.
}
\examples{
# Import data from "/path/on/disk/data.csv to the current working directory
import_data_from_disk("/path/on/disk/data.csv")

# Import data from "/path/on/disk/data.csv" to a custom target path "data/data.csv"
import_data_from_disk("/path/on/disk/data.csv", "data/data.csv")

}
\seealso{
Use \code{\link{digest}} function from the \code{digest} package to compute hash values
for file comparison.
}
\keyword{copy}
\keyword{data}
\keyword{disk}
\keyword{file}
\keyword{import}
